feat(utils): add utils file and move functions over
